backend/pipeline/requirements.txt
kaggle>=1.6.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
